Fix broken hiss-kin cooldown
# Explanation
The type `origins:multiple` gave the hiss-kin a *separate* cooldown. You have one ability in this file that has two abilities:
`rocket_jump` which is ONLY a cooldown, that doesn't apply to the next entry.
The next entry, `rocket_jump` has a cooldown set to 1, which does not match the meter.

# Solution
The multiple types seem unnecessary, so I removed it. At the moment I still have the `origins:multiple` there with one type since for some reason when I remove it, it breaks the entire ability. Until I can find out what I am doing wrong in that regard, this PR remains a draft.